Time: 11 ms (7.09%), Space: 8.5 MB (91.37%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     bool isPalindrome(int x) {
-        double pali=0;
-        double sum=x;;
+        long long int  pali=0;
+        int sum=x;;
        while(x>0){
-        double h=x%10;
+        int h=x%10;
         pali=(pali*10)+h;
         x=x/10;
